refactor haplotype quantification to enable viral quantification
