+++ src/calling/haplotypes/haplotypes.rs
@@ -15,6 +15,11 @@ use rust_htslib::bcf::{
     record::GenotypeAllele::{Phased, Unphased},
     Read,
 };
+
+use good_lp::IntoAffineExpression;
+use good_lp::*;
+use good_lp::{variable, Expression};
+
 use serde::Serialize;
 use serde_json::json;
 use std::collections::{BTreeMap, HashMap};
@@ -245,4 +250,551 @@ impl HaplotypeVariants {
         }
         Ok(haplotype_variants_filtered)
     }
+}
+
+#[derive(Serialize, Debug)]
+pub(crate) struct DatasetVariants {
+    variant: VariantID,
+    vaf: f32,
+}
+#[derive(Serialize, Debug)]
+pub(crate) struct DatasetHaplotypeVariants {
+    variant: VariantID,
+    haplotype: String,
+}
+#[derive(Serialize, Debug)]
+pub(crate) struct DatasetHaplotypeFractions {
+    haplotype: String,
+    fraction: AlleleFreq,
+}
+
+#[derive(Serialize, Debug)]
+pub(crate) struct DatasetAfd {
+    variant: VariantID,
+    allele_freq: AlleleFreq,
+    probability: f64,
+}
+
+pub fn plot_prediction(
+    outdir: &PathBuf,
+    solution: &str,
+    candidate_matrix_values: &Vec<(Vec<VariantStatus>, BitVec)>,
+    haplotypes: &Vec<Haplotype>,
+    variant_calls: &VariantCalls,
+    best_variables: &Vec<f64>,
+) -> Result<()> {
+    let mut file_name = "".to_string();
+    let json = include_str!("../../../templates/prediction.json");
+    let mut blueprint: serde_json::Value = serde_json::from_str(json).unwrap();
+    let mut plot_data_variants = Vec::new();
+    let mut plot_data_haplotype_variants = Vec::new();
+    let mut plot_data_haplotype_fractions = Vec::new();
+    let mut plot_data_covered_variants = Vec::new();
+    let mut plot_data_dataset_afd = Vec::new();
+
+    if &solution == &"lp" {
+        for ((genotype_matrix, coverage_matrix), (variant_id, (af, _))) in
+            candidate_matrix_values.iter().zip(variant_calls.iter())
+        {
+            let mut counter = 0;
+            for (i, _variable) in best_variables.iter().enumerate() {
+                if coverage_matrix[i as u64] {
+                    counter += 1;
+                }
+            }
+            if counter == best_variables.len() {
+                for (i, (variable, haplotype)) in
+                    best_variables.iter().zip(haplotypes.iter()).enumerate()
+                {
+                    if genotype_matrix[i] == VariantStatus::Present {
+                        plot_data_haplotype_fractions.push(DatasetHaplotypeFractions {
+                            haplotype: haplotype.to_string(),
+                            fraction: NotNan::new(*variable).unwrap(),
+                        });
+                        plot_data_haplotype_variants.push(DatasetHaplotypeVariants {
+                            variant: *variant_id,
+                            haplotype: haplotype.to_string(),
+                        });
+                        plot_data_variants.push(DatasetVariants {
+                            variant: *variant_id,
+                            vaf: af.clone(),
+                        });
+                    }
+                }
+            }
+        }
+        file_name.push_str("lp_solution.json");
+    } else if &solution == &"final" {
+        candidate_matrix_values
+            .iter()
+            .zip(variant_calls.iter())
+            .for_each(|((genotypes, covered), (variant_id, (af, afd)))| {
+                best_variables
+                    .iter()
+                    .zip(haplotypes.iter())
+                    .enumerate()
+                    .for_each(|(i, (fraction, haplotype))| {
+                        if genotypes[i] == VariantStatus::Present && covered[i as u64] {
+                            plot_data_haplotype_fractions.push(DatasetHaplotypeFractions {
+                                haplotype: haplotype.to_string(),
+                                fraction: NotNan::new(*fraction).unwrap(),
+                            });
+                            plot_data_haplotype_variants.push(DatasetHaplotypeVariants {
+                                variant: *variant_id,
+                                haplotype: haplotype.to_string(),
+                            });
+                            plot_data_variants.push(DatasetVariants {
+                                variant: *variant_id,
+                                vaf: *af,
+                            });
+                            //addition of one more rect plot for coverage matrix in addition to genotype matrix
+                            //create the plot_data_covered_variants using only the variants that have GT:1 for at least one haplotype.
+                            for (j, haplotype) in haplotypes.iter().enumerate() {
+                                if covered[j as u64] {
+                                    plot_data_covered_variants.push(DatasetHaplotypeVariants {
+                                        variant: *variant_id,
+                                        haplotype: haplotype.to_string(),
+                                    });
+                                }
+                            }
+
+                            //also add the heatmap for afd below the covered panels
+                            for (allele_freq, prob) in afd.iter() {
+                                plot_data_dataset_afd.push(DatasetAfd {
+                                    variant: *variant_id,
+                                    allele_freq: *allele_freq,
+                                    probability: prob.clone(),
+                                })
+                            }
+                        }
+                    });
+            });
+        file_name.push_str("final_solution.json");
+    }
+    let plot_data_variants = json!(plot_data_variants);
+    let plot_data_haplotype_variants = json!(plot_data_haplotype_variants);
+    let plot_data_haplotype_fractions = json!(plot_data_haplotype_fractions);
+    let plot_data_covered_variants = json!(plot_data_covered_variants);
+    let plot_data_dataset_afd = json!(plot_data_dataset_afd);
+
+    blueprint["datasets"]["variants"] = plot_data_variants;
+    blueprint["datasets"]["haplotype_variants"] = plot_data_haplotype_variants;
+    blueprint["datasets"]["haplotype_fractions"] = plot_data_haplotype_fractions;
+    blueprint["datasets"]["covered_variants"] = plot_data_covered_variants;
+    blueprint["datasets"]["allele_frequency_distribution"] = plot_data_dataset_afd;
+
+    let mut parent = outdir.clone();
+    parent.pop();
+    fs::create_dir_all(&parent)?;
+    let file = fs::File::create(parent.join(file_name)).unwrap();
+    serde_json::to_writer(file, &blueprint)?;
+    Ok(())
+}
+
+pub fn linear_program(
+    outdir: &PathBuf,
+    candidate_matrix: &CandidateMatrix,
+    haplotypes: &Vec<Haplotype>,
+    variant_calls: &VariantCalls,
+    lp_cutoff: f64,
+) -> Result<Vec<Haplotype>> {
+    //first init the problem
+    let mut problem = ProblemVariables::new();
+    //introduce variables
+    let variables: Vec<Variable> =
+        problem.add_vector(variable().min(0.0).max(1.0), haplotypes.len());
+
+    //init the constraints
+    let mut constraints: Vec<Expression> = Vec::new();
+
+    //execute the following function to fill up the constraints and create a haplotype_dict
+    let haplotype_dict = collect_constraints_and_variants(
+        candidate_matrix,
+        haplotypes,
+        variant_calls,
+        &variables,
+        &mut constraints,
+    )
+    .unwrap();
+
+    //define temporary variables
+    let t_vars: Vec<Variable> = problem.add_vector(variable().min(0.0).max(1.0), constraints.len());
+
+    //create the model to minimise the sum of temporary variables
+    let mut sum_tvars = Expression::from_other_affine(0.);
+    for t_var in t_vars.iter() {
+        sum_tvars += t_var.into_expression();
+    }
+    let mut model = problem.minimise(sum_tvars.clone()).using(default_solver); // multiple solvers available
+
+    //add a constraint to make sure variables sum up to 1.0.
+    let mut sum = Expression::from_other_affine(0.);
+    for var in variables.iter() {
+        sum += Expression::from_other_affine(var);
+    }
+    model = model.with(constraint!(sum == 1.0));
+
+    //add the constraints to the model
+    for (c, t_var) in constraints.iter().zip(t_vars.iter()) {
+        model = model.with(constraint!(t_var >= c.clone()));
+        model = model.with(constraint!(t_var >= -c.clone()));
+    }
+    // dbg!(&constraints);
+
+    //solve the problem with the default solver, i.e. coin_cbc
+    let solution = model.solve().unwrap();
+
+    let mut best_variables = Vec::new();
+    //finally, print the variables and the sum
+    let mut lp_haplotypes = BTreeMap::new();
+    for (i, (var, haplotype)) in variables.iter().zip(haplotypes.iter()).enumerate() {
+        println!("v{}, {}={}", i, haplotype.to_string(), solution.value(*var));
+        best_variables.push(solution.value(var.clone()).clone());
+        if solution.value(*var) >= lp_cutoff {
+            //the speed of fraction exploration is managable in case of diploid priors
+            lp_haplotypes.insert(haplotype.clone(), solution.value(*var).clone());
+        }
+    }
+    println!("sum = {}", solution.eval(sum_tvars));
+
+    //plot the best result
+    // dbg!(&best_variables.len());
+    let candidate_matrix_values: Vec<(Vec<VariantStatus>, BitVec)> =
+        candidate_matrix.values().cloned().collect();
+    plot_prediction(
+        outdir,
+        &"lp",
+        &candidate_matrix_values,
+        &haplotypes,
+        &variant_calls,
+        &best_variables,
+    )?;
+
+    //extend haplotypes found by linear program, add haplotypes that have the same variants to the final list
+    //and optionally, sort by hamming distance, take the closest x additional alleles according to 'permitted'
+    let mut extended_haplotypes = Vec::new();
+    lp_haplotypes.iter().for_each(|(f_haplotype, _)| {
+        let variants = haplotype_dict.get(&f_haplotype).unwrap().clone();
+        haplotype_dict
+            .iter()
+            .for_each(|(haplotype, haplotype_variants)| {
+                if &variants == haplotype_variants && !extended_haplotypes.contains(haplotype) {
+                    //fix: the last operand '&&' is required to avoid duplicate additions
+                    extended_haplotypes.push(haplotype.clone());
+                } else {
+                    let permitted: i64 = 3; //this number should be discussed.
+                    let mut difference = vec![];
+                    for i in haplotype_variants.iter() {
+                        if !variants.contains(&i) {
+                            difference.push(i);
+                        }
+                    }
+                    if (difference.len() as i64 <= permitted)
+                        && ((variants.len() as i64 - haplotype_variants.len() as i64).abs()
+                            <= permitted)
+                        && !extended_haplotypes.contains(&haplotype)
+                    //fix: the last operand '&&' is required to avoid duplicate additions
+                    {
+                        extended_haplotypes.push(haplotype.clone());
+                    }
+                }
+            });
+    });
+    dbg!(&lp_haplotypes);
+    //diploid-subclonal max N haplotypes
+    // let max_haplotypes = 5;
+    let lp_keys: Vec<_> = lp_haplotypes.keys().cloned().collect();
+    let lp_values_original: Vec<_> = lp_haplotypes.values().cloned().collect();
+    let mut lp_values: Vec<_> = lp_haplotypes.values().cloned().collect();
+    lp_values.sort_by(|a, b| OrderedFloat(*b).cmp(&OrderedFloat(*a)));
+    let mut selected_haplotypes = Vec::new();
+    for x in lp_values.iter() {
+        for (i, y) in lp_values_original.iter().enumerate() {
+            if x == y {
+                selected_haplotypes.push(lp_keys[i].clone());
+            }
+        }
+    }
+    // if max_haplotypes < selected_haplotypes.len() {
+    //     selected_haplotypes = selected_haplotypes[0..max_haplotypes].to_vec();
+    // }
+    // dbg!(&selected_haplotypes);
+
+    dbg!(&extended_haplotypes);
+    Ok(extended_haplotypes)
+}
+
+pub fn write_results(
+    outdir: &PathBuf,
+    data: &Data,
+    event_posteriors: &Vec<(HaplotypeFractions, LogProb)>,
+    final_haplotypes: &Vec<Haplotype>,
+    _prior: String,
+    variant_info: bool,
+) -> Result<()> {
+    //firstly add variant query and probabilities to the outout table for each event
+    let variant_calls: Vec<AlleleFreqDist> = data
+        .variant_calls
+        .iter()
+        .map(|(_, (_, afd))| afd.clone())
+        .collect();
+    let mut event_queries: Vec<BTreeMap<VariantID, (AlleleFreq, LogProb)>> = Vec::new();
+    // let event_posteriors = computed_model.event_posteriors();
+    if variant_info {
+        event_posteriors.iter().for_each(|(fractions, _)| {
+            let mut vaf_queries: BTreeMap<VariantID, (AlleleFreq, LogProb)> = BTreeMap::new();
+            data.candidate_matrix
+                .iter()
+                .zip(variant_calls.iter())
+                .for_each(|((variant_id, (genotypes, covered)), afd)| {
+                    let mut denom = NotNan::new(1.0).unwrap();
+                    let mut vaf_sum = NotNan::new(0.0).unwrap();
+                    let mut counter = 0;
+                    fractions.iter().enumerate().for_each(|(i, fraction)| {
+                        if genotypes[i] == VariantStatus::Present && covered[i as u64] {
+                            vaf_sum += *fraction;
+                            counter += 1;
+                        } else if genotypes[i] == VariantStatus::Unknown && covered[i as u64] {
+                            todo!();
+                        } else if genotypes[i] == VariantStatus::Unknown
+                            && covered[i as u64] == false
+                        {
+                            todo!();
+                        } else if genotypes[i] == VariantStatus::NotPresent
+                            && covered[i as u64] == false
+                        {
+                            denom -= *fraction;
+                        }
+                    });
+                    if denom > NotNan::new(0.0).unwrap() {
+                        vaf_sum /= denom;
+                    }
+                    vaf_sum = NotNan::new((vaf_sum * NotNan::new(100.0).unwrap()).round()).unwrap()
+                        / NotNan::new(100.0).unwrap();
+                    if !afd.is_empty() && counter > 0 {
+                        let answer = afd.vaf_query(&vaf_sum);
+                        vaf_queries.insert(*variant_id, (vaf_sum, answer.unwrap()));
+                    } else {
+                        ()
+                    }
+                });
+            event_queries.push(vaf_queries);
+        });
+    }
+    // Then,print TSV table with results
+    // Columns: posterior_prob, haplotype_a, haplotype_b, haplotype_c, ...
+    // with each column after the first showing the fraction of the respective haplotype
+    let mut wtr = csv::Writer::from_path(outdir)?;
+    let mut headers: Vec<_> = vec!["density".to_string(), "odds".to_string()];
+    let haplotypes_str: Vec<String> = final_haplotypes
+        .clone()
+        .iter()
+        .map(|h| h.to_string())
+        .collect();
+    headers.extend(haplotypes_str);
+    if variant_info {
+        let variant_names = event_queries[0]
+            .keys()
+            .map(|key| format!("{:?}", key))
+            .collect::<Vec<String>>();
+        headers.extend(variant_names); //add variant names as separate columns
+    }
+    wtr.write_record(&headers)?;
+
+    //write best record on top
+    let mut records = Vec::new();
+    // let mut event_posteriors = computed_model.event_posteriors(); //compute a second time because event_posteriors can't be cloned from above.
+    let (haplotype_frequencies, best_density) = event_posteriors.iter().next().unwrap();
+    let best_odds = 1;
+    let format_f64 = |number: f64, records: &mut Vec<String>| {
+        if number <= 0.01 {
+            records.push(format!("{:+.2e}", number))
+        } else {
+            records.push(format!("{:.2}", number))
+        }
+    };
+    format_f64(best_density.exp(), &mut records);
+    records.push(best_odds.to_string());
+    let format_freqs = |frequency: NotNan<f64>, records: &mut Vec<String>| {
+        if frequency <= NotNan::new(0.01).unwrap() {
+            records.push(format!("{:+.2e}", NotNan::into_inner(frequency)))
+        } else {
+            records.push(format!("{:.2}", frequency))
+        }
+    };
+    haplotype_frequencies
+        .iter()
+        .for_each(|frequency| format_freqs(*frequency, &mut records));
+
+    if variant_info {
+        //add vaf queries and probabilities for the first event to the output table
+        let queries: Vec<(AlleleFreq, LogProb)> = event_queries
+            .iter()
+            .next()
+            .unwrap()
+            .values()
+            .cloned()
+            .collect();
+        queries.iter().for_each(|(query, answer)| {
+            let prob = f64::from(Prob::from(*answer));
+            if prob <= 0.01 {
+                records.push(format!("{}{}{:+.2e}", query, ":", prob));
+            } else {
+                records.push(format!("{}{}{:.2}", query, ":", prob));
+            }
+        });
+    }
+
+    wtr.write_record(records)?;
+
+    //write the rest of the records
+    dbg!(&event_posteriors);
+
+    if variant_info {
+        event_posteriors
+            .iter()
+            .skip(1)
+            .zip(event_queries.iter().skip(1))
+            .for_each(|((haplotype_frequencies, density), queries)| {
+                let mut records = Vec::new();
+                let odds = (density - best_density).exp();
+                format_f64(density.exp(), &mut records);
+                format_f64(odds, &mut records);
+                haplotype_frequencies
+                    .iter()
+                    .for_each(|frequency| format_freqs(*frequency, &mut records));
+
+                queries.iter().for_each(|(_, (query, answer))| {
+                    let prob = f64::from(Prob::from(*answer));
+                    if prob <= 0.01 {
+                        records.push(format!("{}{}{:+.2e}", query, ":", prob));
+                    } else {
+                        records.push(format!("{}{}{:.2}", query, ":", prob));
+                    }
+                });
+                wtr.write_record(records).unwrap();
+            });
+    } else {
+        event_posteriors
+            .iter()
+            .skip(1)
+            .for_each(|(haplotype_frequencies, density)| {
+                let mut records = Vec::new();
+                let odds = (density - best_density).exp();
+                format_f64(density.exp(), &mut records);
+                format_f64(odds, &mut records);
+                haplotype_frequencies
+                    .iter()
+                    .for_each(|frequency| format_freqs(*frequency, &mut records));
+
+                wtr.write_record(records).unwrap();
+            });
+    }
+
+    Ok(())
+}
+
+pub fn collect_constraints_and_variants(
+    candidate_matrix: &CandidateMatrix,
+    haplotypes: &Vec<Haplotype>,
+    variant_calls: &VariantCalls,
+    variables: &Vec<Variable>,
+    constraints: &mut Vec<Expression>,
+) -> Result<HashMap<Haplotype, Vec<VariantID>>> {
+    let candidate_matrix_values: Vec<(Vec<VariantStatus>, BitVec)> =
+        candidate_matrix.values().cloned().collect();
+    //collect haplotype-to-variants information
+    let mut haplotype_dict: HashMap<Haplotype, Vec<VariantID>> =
+        haplotypes.iter().map(|h| (h.clone(), vec![])).collect();
+    //variant-wise iteration
+    let mut expr = Expression::from_other_affine(0.); // A constant expression
+    for ((genotype_matrix, coverage_matrix), (variant, (af, _))) in
+        candidate_matrix_values.iter().zip(variant_calls.iter())
+    {
+        let mut fraction_cont = Expression::from_other_affine(0.);
+        let _prime_fraction_cont = Expression::from_other_affine(0.);
+        let _vaf = Expression::from_other_affine(0.);
+        let mut counter = 0;
+        for (i, _variable) in variables.iter().enumerate() {
+            if coverage_matrix[i as u64] {
+                counter += 1;
+            }
+        }
+        if counter == variables.len() {
+            for (i, (variable, haplotype)) in variables.iter().zip(haplotypes.iter()).enumerate() {
+                if genotype_matrix[i] == VariantStatus::Present {
+                    fraction_cont += *variable;
+                    let mut existing = haplotype_dict.get(&haplotype).unwrap().clone();
+                    existing.push(variant.clone());
+                    haplotype_dict.insert(haplotype.clone(), existing);
+                }
+            }
+            let expr_to_add = fraction_cont - af.clone().into_expression();
+            constraints.push(expr_to_add.clone());
+            expr += expr_to_add;
+        }
+    }
+    Ok(haplotype_dict)
+}
+
+#[derive(Serialize, Debug)]
+pub(crate) struct DatasetHaplotypeFractionsWsolution {
+    haplotype: String,
+    fraction: AlleleFreq,
+    solution_number: usize,
+}
+
+#[derive(Serialize, Debug)]
+pub(crate) struct DatasetDensitySolution {
+    density: f64,
+    solution_number: usize,
+}
+
+pub fn plot_densities(
+    outdir: &PathBuf,
+    event_posteriors: &Vec<(HaplotypeFractions, LogProb)>,
+    final_haplotypes: &Vec<Haplotype>,
+    file_prefix: &str,
+) -> Result<()> {
+    let file_name = format!("{}_solutions.json", file_prefix.to_string());
+    let json = include_str!("../../../templates/densities.json");
+    let mut blueprint: serde_json::Value = serde_json::from_str(json).unwrap();
+    let mut plot_data_fractions = Vec::new();
+    let mut plot_density = Vec::new();
+
+    //take first 10 solutions, if the events are less than that, then take the length
+    let plot_first_events = 10;
+    let num_events = event_posteriors.len();
+    let mut new_event_posteriors = event_posteriors.clone();
+    if num_events > plot_first_events {
+        new_event_posteriors = new_event_posteriors[0..plot_first_events].to_vec();
+    }
+    for (i, (fractions, logprob)) in new_event_posteriors.iter().enumerate() {
+        plot_density.push(DatasetDensitySolution {
+            density: logprob.exp().clone(),
+            solution_number: i,
+        });
+        for (f, h) in fractions.iter().zip(final_haplotypes.iter()) {
+            //fractions and final haplotypes are already ordered.
+            if f != &NotNan::new(0.0).unwrap() {
+                plot_data_fractions.push(DatasetHaplotypeFractionsWsolution {
+                    haplotype: h.to_string(),
+                    fraction: f.clone(),
+                    solution_number: i,
+                });
+            }
+        }
+    }
+
+    //write to json
+    let plot_data_fractions = json!(plot_data_fractions);
+    let plot_density = json!(plot_density);
+    blueprint["datasets"]["haplotype_fractions"] = plot_data_fractions;
+    blueprint["datasets"]["densities"] = plot_density;
+
+    let mut parent = outdir.clone();
+    parent.pop();
+    let file = fs::File::create(parent.join(file_name)).unwrap();
+    serde_json::to_writer(file, &blueprint)?;
+    Ok(())
 }
